make settlement target circles interactive
#7
